Move `ShutdownHook` class from `operator-common` to `cluster-operator` (#9877)
