+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/ShutdownHook.java
@@ -2,7 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.operator.common;
+package io.strimzi.operator.cluster;
 
 import io.vertx.core.Vertx;
 import org.apache.logging.log4j.LogManager;
